+++ problems/resilient_scraper/sandbox/environment.py
@@ -109,7 +109,7 @@ class ResilientScraperEnvironment:
     # Scoring
     # ------------------------------------------------------------------
 
-    def compute_score(self) -> int:
+    def compute_score(self) -> float:
         return 100 if self._success else 0
 
     # ------------------------------------------------------------------
